requirements.txt
nltk==3.3
GitPython==2.1.11